Patch by Daniel Furrer: - fixed lots of other small beos bugs - commented out a lot of beos-code which needs more investigation/a rewrite to get the thing to compile a bit further - continued the seperation of .._BeOSFrame/BeOSFrameImpl
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@13408 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/ap/beos/ap_BeOSDialog_Paragraph.cpp
@@ -26,9 +26,10 @@
 #include "ut_debugmsg.h"
 #include "ut_types.h"
 
+#include <InterfaceKit.h>
+
 #include "xap_App.h"
 #include "xap_BeOSApp.h"
-#include "xap_BeOSFrame.h"
 
 #include "ap_Dialog_Id.h"
 
@@ -248,8 +249,8 @@ void ParagraphWin::SetDlg(AP_BeOSDialog_Paragraph *brk)
 	view1->RemoveSelf();
 	view2->RemoveSelf();
 	
-	BTab* tab1 = pView->RemoveTab(0);
-	BTab* tab2 = pView->RemoveTab(0);
+	BTab* tab1 = pView->RemoveTab((long int)0);
+	BTab* tab2 = pView->RemoveTab((long int)0);
 	
 	
 	pNewTabView->AddTab(view1 , tab1);
